requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.12
tavily-python==0.5.0
yt-dlp>=2024.12.23
openai-whisper==20240930
anthropic==0.40.0
aiofiles==24.1.0
